+++ examplepatches/humzakh/HK-Morphe-Patches/patches/src/main/kotlin/app/template/patches/reddit/customclients/sync/syncforreddit/Constants.kt
@@ -9,22 +9,26 @@ internal val SyncForRedditFreeCompatible = Compatibility(
     targets = listOf(AppTarget(version = "v23.06.30-13:39"))
 )
 
-internal val SyncForLemmyCompatible = Compatibility(
-    name = "Sync for Lemmy",
-    packageName = "io.syncapps.lemmy_sync",
-    targets = listOf(AppTarget(version = null))
+internal val SyncForRedditProCompatible = Compatibility(
+    name = "Sync for Reddit Pro",
+    packageName = "com.laurencedawson.reddit_sync.pro",
+    targets = listOf(AppTarget(version = "v23.06.30-13:39"))
+)
+
+internal val SyncForRedditDevCompatible = Compatibility(
+    name = "Sync for Reddit Dev",
+    packageName = "com.laurencedawson.reddit_sync.dev",
+    targets = listOf(AppTarget(version = "v23.06.30-13:39"))
 )
 
 internal val SyncForRedditCompatible = arrayOf(
     SyncForRedditFreeCompatible,
-    Compatibility(
-        name = "Sync for Reddit Pro",
-        packageName = "com.laurencedawson.reddit_sync.pro",
-        targets = listOf(AppTarget(version = null))
-    ),
-    Compatibility(
-        name = "Sync for Reddit Dev",
-        packageName = "com.laurencedawson.reddit_sync.dev",
-        targets = listOf(AppTarget(version = null))
-    )
+    SyncForRedditProCompatible,
+    SyncForRedditDevCompatible
+)
+
+internal val SyncForLemmyCompatible = Compatibility(
+    name = "Sync for Lemmy",
+    packageName = "io.syncapps.lemmy_sync",
+    targets = listOf(AppTarget(version = null))
 )